Fix conde based on comment 2
